requirements.txt
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
pytz==2018.4
